Move bertext to extractor/__init__.py
